Use local directory instead of home directory

+++ driver.py
@@ -4,6 +4,7 @@ import sys
 import argparse
 import os
 import random
+import tempfile
 
 from pathlib import Path
 from pem import *
@@ -13,7 +14,7 @@ from verifySignature import *
 def main():
 
     ### command-line argument processing
-    # usage: ./armor-driver [-h] [--chain INPUT] [--trust_store CA_STORE] [--purpose CHECK_PURPOSE] 
+    # usage: ./armor-driver [-h] [--chain INPUT] [--trust_store CA_STORE] [--purpose CHECK_PURPOSE]
     parser = argparse.ArgumentParser(description='ARMOR command-line arguments')
     parser.add_argument('--chain', type=str,
                         help='Input certificate chain location')
@@ -53,48 +54,51 @@ def main():
 
     ep = random.random()
     args = sys.argv
-    home_dir = str(Path.home())
-    filename_certchain = input_chain
-    filename_aeres_output = home_dir + "/.residuals/temp_{}.txt".format(ep)
+    # home_dir = str(Path.home())
+    script_dir = os.path.dirname(os.path.realpath(__file__))
 
-    if not os.path.exists(home_dir + "/.residuals/"):
-        os.mkdir(home_dir + "/.residuals/")
+    with tempfile.TemporaryDirectory() as tmp_dir:
+        filename_certchain = input_chain
+        filename_aeres_output = tmp_dir + "/.residuals/temp_{}.txt".format(ep)
 
-    if input_chain.endswith(".der"):
-        if input_purpose == None:
-            cmd = ['{}/.armor/armor-bin --DER {} {} > {}'.format(home_dir, filename_certchain, input_CA_store, filename_aeres_output)]
+        if not os.path.exists(tmp_dir + "/.residuals/"):
+            os.mkdir(tmp_dir + "/.residuals/")
+
+        if input_chain.endswith(".der"):
+            if input_purpose == None:
+                cmd = ['{}/armor-bin --DER {} {} > {}'.format(script_dir, filename_certchain, input_CA_store, filename_aeres_output)]
+            else:
+                cmd = ['{}/armor-bin --DER --purpose {} {} {} > {}'.format(script_dir, input_purpose, filename_certchain, input_CA_store, filename_aeres_output)]
+        else: ## for .pem and .crt
+            if input_purpose == None:
+                cmd = ['{}/armor-bin {} {} > {}'.format(script_dir, filename_certchain, input_CA_store, filename_aeres_output)]
+            else:
+                cmd = ['{}/armor-bin --purpose {} {} {} > {}'.format(script_dir, input_purpose, filename_certchain, input_CA_store, filename_aeres_output)]
+
+        aeres_res = subprocess.getoutput(cmd)
+        print(aeres_res)
+
+        if aeres_res.__contains__("failed") or aeres_res.__contains__("error") or aeres_res.__contains__("Error") \
+                or aeres_res.__contains__("exception") or aeres_res.__contains__("TLV: cert") \
+                or aeres_res.__contains__("cannot execute binary file") or aeres_res.__contains__("more bytes remain") \
+                or aeres_res.__contains__("incomplete read") or aeres_res.__contains__("not found"):
+            print("AERES syntactic or semantic checks: failed")
+            os.remove(filename_aeres_output)
+            return False
         else:
-            cmd = ['{}/.armor/armor-bin --DER --purpose {} {} {} > {}'.format(home_dir, input_purpose, filename_certchain, input_CA_store, filename_aeres_output)]
-    else: ## for .pem and .crt
-        if input_purpose == None:
-            cmd = ['{}/.armor/armor-bin {} {} > {}'.format(home_dir, filename_certchain, input_CA_store, filename_aeres_output)]
-        else:
-            cmd = ['{}/.armor/armor-bin --purpose {} {} {} > {}'.format(home_dir, input_purpose, filename_certchain, input_CA_store, filename_aeres_output)]
+            print("AERES syntactic and semantic checks: passed")
 
-    aeres_res = subprocess.getoutput(cmd)
-    print(aeres_res)
-
-    if aeres_res.__contains__("failed") or aeres_res.__contains__("error") or aeres_res.__contains__("Error") \
-            or aeres_res.__contains__("exception") or aeres_res.__contains__("TLV: cert") \
-            or aeres_res.__contains__("cannot execute binary file") or aeres_res.__contains__("more bytes remain") \
-            or aeres_res.__contains__("incomplete read") or aeres_res.__contains__("not found"):
-        print("AERES syntactic or semantic checks: failed")
+        readData(filename_aeres_output)
         os.remove(filename_aeres_output)
-        return False
-    else:
-        print("AERES syntactic and semantic checks: passed")
 
-    readData(filename_aeres_output)
-    os.remove(filename_aeres_output)
+        sign_verify_res = verifySignatures()
+        if sign_verify_res == "false":
+            print("Signature verification: failed")
+            return False
+        else:
+            print("Signature verification: passed")
 
-    sign_verify_res = verifySignatures()
-    if sign_verify_res == "false":
-        print("Signature verification: failed")
-        return False
-    else:
-        print("Signature verification: passed")
-
-    return True
+        return True
 
 if __name__ == "__main__":
     res = main()
